Add reproduction setup
It takes a bit of a non-minimal setup to reproduce.

It seems that application-2, application-3, and application-4 all wait
for application-1 to finish its build. It's unclear why this is
happening, but it seems to consistently happen no matter what.

We add the profile so it is a bit easier to see the issue without having
to actually run everything.